% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_n_boxplot.R
\name{top_n_boxplot}
\alias{top_n_boxplot}
\title{Show top N values of each group on the boxplot}
\usage{
top_n_boxplot(dataset, value, group, n = 10)
}
\arguments{
\item{dataset}{Dataframe or tibble with at least two columns}

\item{value}{A numeric column that is used for ordering}

\item{group}{A factor or character column used for grouping}

\item{n}{An integer showing the number of top values to be plotted. Default value is 10.}
}
\value{
A ggplot object, visualization with two layers (boxplot and scatterplot)
}
\description{
This function adds top N highest values of every group to the boxplot
visualization
}
\details{
Function would not work if any of the groups has less than n observations
}
\examples{
mtcars \%>\%
dplyr::mutate(am = as.factor(am)) \%>\%
top_n_boxplot(mpg, am, 5)
}
